add-some-date-to-DB

+++ ElectionLand/Models/User.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -8,6 +9,7 @@ namespace ElectionLand.Models
     
     public class User
     {
+
         public int Id { get; set; }
 
         public string FirstName { get; set; }
@@ -16,7 +18,7 @@ namespace ElectionLand.Models
 
         public DateTime BirthDate { get; set; }
 
-        public int PIN { get; set; }
+        public long PIN { get; set; }
 
         public string Email { get; set; }
